pyproject.toml
[tool.poetry]
name = "pynamodb-factoryboy"
version = "0.1.0"
description = ""
authors = ["Nikita Antonenkov <krigar1184@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytest-localstack = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "6.2.2"
factory-boy = "^3.2.0"
pynamodb = "^5.0.3"
pytest-factoryboy = "^2.1.0"
pdbpp = "^0.10.2"
docker = "^4.4.4"
pytest-env = "^0.6.2"
tox = "^3.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
